+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] 2025-07-15 Bug Fixes and Improvements
+
+### Added
+
+- Library version from CMake projecto to `ArffFiles.hpp`
+- Library `catch2` as a conan test requirement
+- Install target for CMake
+
 ## [1.2.0] 2025-06-27 Refactoring and Improvements
 
 ### Added